better offset calculation for decorators; separate ScrollDiffDecorator
